:sparkles: Add SentEval CLI
